+++ src/main/java/sopt/makers/jwt/verifier/JwtJwkVerifierApplication.java
@@ -3,7 +3,7 @@ package sopt.makers.jwt.verifier;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.boot.context.properties.EnableConfigurationProperties;
-import sopt.makers.jwt.verifier.config.property.AuthClientProperty;
+import sopt.makers.jwt.verifier.external.auth.AuthClientProperty;
 
 @SpringBootApplication
 @EnableConfigurationProperties(AuthClientProperty.class)
