+++ src/main/java/sopt/makers/jwt/verifier/common/exception/BaseException.java
