+++ src/main/java/sopt/makers/jwt/verifier/common/response/BaseResponse.java
@@ -1,4 +1,4 @@
-package sopt.makers.jwt.verifier.api;
+package sopt.makers.jwt.verifier.common.response;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import lombok.Builder;
