--- src/main/java/sopt/makers/jwt/verifier/exception/JwkException.java
@@ -1,9 +1,0 @@
-package sopt.makers.jwt.verifier.exception;
-
-import sopt.makers.jwt.verifier.code.failure.JwkFailure;
-
-public class JwkException extends BaseException {
-    public JwkException(JwkFailure failure) {
-        super(failure);
-    }
-}
