--- src/main/java/sopt/makers/jwt/verifier/exception/JwtException.java
@@ -1,9 +1,0 @@
-package sopt.makers.jwt.verifier.exception;
-
-import sopt.makers.jwt.verifier.code.failure.JwtFailure;
-
-public class JwtException extends BaseException {
-    public JwtException(JwtFailure failure) {
-        super(failure);
-    }
-}
