+++ src/main/java/sopt/makers/jwt/verifier/jwt/exception/JwtException.java
@@ -1,0 +1,10 @@
+package sopt.makers.jwt.verifier.jwt.exception;
+
+import sopt.makers.jwt.verifier.jwt.code.JwtFailure;
+import sopt.makers.jwt.verifier.exception.BaseException;
+
+public class JwtException extends BaseException {
+    public JwtException(JwtFailure failure) {
+        super(failure);
+    }
+}
