+++ src/main/java/sopt/makers/jwt/verifier/security/filter/JwtAuthenticationExceptionFilter.java
@@ -7,8 +7,8 @@ import jakarta.servlet.http.HttpServletResponse;
 import org.springframework.lang.NonNull;
 import org.springframework.stereotype.Component;
 import org.springframework.web.filter.OncePerRequestFilter;
-import sopt.makers.jwt.verifier.exception.JwtException;
-import sopt.makers.jwt.verifier.util.ResponseUtil;
+import sopt.makers.jwt.verifier.jwt.exception.JwtException;
+import sopt.makers.jwt.verifier.common.util.ResponseUtil;
 
 import java.io.IOException;
